pyproject.toml
[project]
name = "taxer-qa-automation"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "black>=26.3.1",
    "pytest>=9.0.2",
    "pytest-playwright>=0.7.2",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
]
